backend/requirements.txt
# KAVACH-AI backend requirements
# Install from project root: pip install -r requirements.txt

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic-settings>=2.4.0
aiofiles>=24.1.0

torch>=2.4.0
torchvision>=0.19.0
timm>=1.0.9
transformers>=4.45.0

onnxruntime>=1.18.0

opencv-python-headless>=4.10.0.84
pillow>=10.4.0

soundfile>=0.12.1
librosa>=0.10.2.post1
scipy>=1.14.1

numpy>=1.26.4
scikit-learn>=1.5.0

pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0

pyyaml>=6.0.2
